+++ Sources/PayPalWebPayments/PayPalVaultRequest.swift
@@ -3,8 +3,9 @@ import Foundation
 /// A request to vault a PayPal payment method
 public struct PayPalVaultRequest {
 
+    // NEXT_MAJOR_VERSION: - Remove URL property
     /// PayPal approval URL returned as the `href` from the setup token API call
-    public let url: URL
+    public let url: URL? = nil
 
     /// ID for the setup token associated with the vault
     /// Returned as  top level `id` from the setup token API call
@@ -14,8 +15,14 @@ public struct PayPalVaultRequest {
     /// - Parameters:
     ///    - url: PayPal approval URL returned as the `href` from the setup token API call
     ///    - setupTokenID: An ID for the setup token associated with the vault
+    @available(*, deprecated, message: "Use `init(setupTokenID:)` instead.")
     public init(url: URL, setupTokenID: String) {
-        self.url = url
+        self.setupTokenID = setupTokenID
+    }
+    
+    /// Creates an instance of a PayPal vault request
+    /// - Parameter setupTokenID: An ID for the setup token associated with the vault
+    public init(setupTokenID: String) {
         self.setupTokenID = setupTokenID
     }
 }